+++ packages/charting/src/components/Legends/Legends.types.ts
@@ -48,7 +48,17 @@ export interface ILegend {
   /**
    * Defines the function that is executed on clicking this legend
    */
-  action: VoidFunction;
+  action?: VoidFunction;
+
+  /**
+   * Defines the function that is executed upon hovering over the legend
+   */
+  hoverAction?: VoidFunction;
+
+  /**
+   * Defines the function that is executed upon moving the mouse away from the legend
+   */
+  onMouseOutAction?: VoidFunction;
 
   /**
    * The color for the legend