+++ packages/charting/src/types/IDataPoint.ts
@@ -11,3 +11,37 @@ export interface IDataPoint {
    */
   y: number;
 }
+
+export interface IChartDataPoint {
+  /**
+   * Legend text for the datapoint in the chart
+   */
+  legend?: string;
+
+  /**
+   * data the datapoint in the chart
+   */
+  data?: number;
+
+  /**
+   * onClick action for each datapoint in the chart
+   */
+  onClick?: VoidFunction;
+
+  /**
+   * color for the legend in the chart
+   */
+  color?: string;
+}
+
+export interface IChartProps {
+  /**
+   * chart title for the chart
+   */
+  chartTitle?: string;
+
+  /**
+   * data for the points in the chart
+   */
+  chartData?: IChartDataPoint[];
+}