+++ packages/dashboard/src/index.ts
@@ -3,3 +3,4 @@ export * from './components/Card/index';
 export * from './components/Nav/index';
 export * from './components/Recommendation/index';
 export * from './components/MultiCount/index';
+export * from './components/Section/index';