+++ packages/experiments/src/components/Stack/Stack.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { createStatelessComponent, IStyleableComponentProps, IViewComponentProps } from '../../Foundation';
+import { createStatelessComponent, IStyleableComponent, IViewComponentProps } from '../../Foundation';
 import StackItem from './StackItem/StackItem';
 import { IStackItemProps, IStackItemStyles } from './StackItem/StackItem.types';
 import { IStackProps, IStackStyles } from './Stack.types';
@@ -7,10 +7,20 @@ import { styles } from './Stack.styles';
 import { mergeStyles } from 'office-ui-fabric-react/lib/Styling';
 
 const StackItemType = (<StackItem /> as React.ReactElement<IStackItemProps> &
-  IStyleableComponentProps<IStackItemProps, IStackItemStyles>).type;
+  IStyleableComponent<IStackItemProps, IStackItemStyles>).type;
 
 const view = (props: IViewComponentProps<IStackProps, IStackStyles>) => {
-  const { renderAs: RootType = 'div', classNames, gap, horizontal, shrinkItems } = props;
+  const {
+    renderAs: RootType = 'div',
+    classNames,
+    gap,
+    horizontal,
+    shrinkItems,
+    verticalAlignment,
+    horizontalAlignment,
+    grow,
+    ...rest
+  } = props;
 
   const stackChildren: (React.ReactChild | null)[] = React.Children.map(
     props.children,
@@ -49,7 +59,11 @@ const view = (props: IViewComponentProps<IStackProps, IStackStyles>) => {
     }
   );
 
-  return <RootType className={classNames.root}>{stackChildren}</RootType>;
+  return (
+    <RootType {...rest} className={classNames.root}>
+      {stackChildren}
+    </RootType>
+  );
 };
 
 const StackStatics = {