+++ packages/office-ui-fabric-react/src/components/Checkbox/Checkbox.test.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 import * as renderer from 'react-test-renderer';
+import { mount } from 'enzyme';
 
 import { Checkbox } from './index';
 
@@ -8,5 +9,9 @@ describe('Checkbox', () => {
     const component = renderer.create(<Checkbox label="Standard checkbox" ariaDescribedBy={'descriptionID'} />);
     const tree = component.toJSON();
     expect(tree).toMatchSnapshot();
+  });
+  it('renders Checkbox with id correctly', () => {
+    const component = mount(<Checkbox label="Standard checkbox" ariaDescribedBy={'descriptionID'} id="my-checkbox" />);
+    expect(component.find('button').prop('id')).toEqual('my-checkbox');
   });
 });