+++ packages/office-ui-fabric-react/src/components/DetailsList/DetailsRow.base.tsx
@@ -18,7 +18,7 @@ import { FocusZone, FocusZoneDirection, IFocusZone } from '../../FocusZone';
 import { SelectionMode, SELECTION_CHANGE } from '../../utilities/selection/interfaces';
 import { CollapseAllVisibility } from '../../GroupedList';
 import { IDragDropOptions } from './../../utilities/dragdrop/interfaces';
-import { IDetailsRowProps } from './DetailsRow.types';
+import { IDetailsRowBaseProps } from './DetailsRow.types';
 import { IDetailsRowCheckProps } from './DetailsRowCheck.types';
 import { IDetailsRowStyleProps, IDetailsRowStyles } from './DetailsRow.types';
 import { classNamesFunction } from '../../Utilities';
@@ -43,7 +43,9 @@ export interface IDetailsRowState {
 
 const DEFAULT_DROPPING_CSS_CLASS = 'is-dropping';
 
-export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowState> {
+const NO_COLUMNS: IColumn[] = [];
+
+export class DetailsRowBase extends BaseComponent<IDetailsRowBaseProps, IDetailsRowState> {
   private _root: HTMLElement | undefined;
   private _cellMeasurer = createRef<HTMLSpanElement>();
   private _focusZone = createRef<IFocusZone>();
@@ -51,7 +53,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowS
   private _hasMounted: boolean;
   private _dragDropSubscription: IDisposable;
 
-  constructor(props: IDetailsRowProps) {
+  constructor(props: IDetailsRowBaseProps) {
     super(props);
 
     this.state = {
@@ -87,7 +89,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowS
     }
   }
 
-  public componentDidUpdate(previousProps: IDetailsRowProps) {
+  public componentDidUpdate(previousProps: IDetailsRowBaseProps) {
     const state = this.state;
     const { item, onDidMount } = this.props;
     const { columnMeasureInfo } = state;
@@ -141,14 +143,14 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowS
     }
   }
 
-  public componentWillReceiveProps(newProps: IDetailsRowProps): void {
+  public componentWillReceiveProps(newProps: IDetailsRowBaseProps): void {
     this.setState({
       selectionState: this._getSelectionState(newProps),
       groupNestingDepth: newProps.groupNestingDepth
     });
   }
 
-  public shouldComponentUpdate(nextProps: IDetailsRowProps, nextState: IDetailsRowState): boolean {
+  public shouldComponentUpdate(nextProps: IDetailsRowBaseProps, nextState: IDetailsRowState): boolean {
     if (this.props.useReducedRowRenderer) {
       if (this.state.selectionState) {
         const newSelectionState = this._getSelectionState(nextProps);
@@ -165,7 +167,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowS
   public render(): JSX.Element {
     const {
       className,
-      columns,
+      columns = NO_COLUMNS,
       dragDropEvents,
       item,
       itemIndex,
@@ -305,7 +307,8 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowS
    * @param {(width: number) => void} onMeasureDone (the call back function when finish measure)
    */
   public measureCell(index: number, onMeasureDone: (width: number) => void): void {
-    const column = assign({}, this.props.columns[index]) as IColumn;
+    const { columns = NO_COLUMNS } = this.props;
+    const column = assign({}, columns[index]) as IColumn;
 
     column.minWidth = 0;
     column.maxWidth = 999999;
@@ -329,7 +332,7 @@ export class DetailsRowBase extends BaseComponent<IDetailsRowProps, IDetailsRowS
     return <DetailsRowCheck {...props} />;
   }
 
-  private _getSelectionState(props: IDetailsRowProps): IDetailsRowSelectionState {
+  private _getSelectionState(props: IDetailsRowBaseProps): IDetailsRowSelectionState {
     const { itemIndex, selection } = props;
 
     return {