+++ packages/utilities/src/customizable.tsx
@@ -1,8 +1,12 @@
 import * as React from 'react';
+import * as PropTypes from 'prop-types';
 import { Customizations } from './Customizations';
 import { hoistStatics } from './hoistStatics';
-import { CustomizerContext, ICustomizerContext } from './Customizer';
 import { concatStyleSets } from '@uifabric/merge-styles';
+
+export const CustomizableContextTypes = {
+  customizations: PropTypes.object
+};
 
 export function customizable(
   scope: string,
@@ -19,9 +23,11 @@ export function customizable(
     const resultClass = class ComponentWithInjectedProps extends React.Component<P, {}> {
       public static displayName: string = 'Customized' + scope;
 
+      public static contextTypes = CustomizableContextTypes;
+
       // tslint:disable-next-line:no-any
-      constructor(props: P) {
-        super(props);
+      constructor(props: P, context: any) {
+        super(props, context);
 
         this._onSettingChanged = this._onSettingChanged.bind(this);
       }
@@ -35,23 +41,17 @@ export function customizable(
       }
 
       public render(): JSX.Element {
-        return (
-          <CustomizerContext.Consumer>
-            {(context: ICustomizerContext) => {
-              const defaultProps = Customizations.getSettings(fields, scope, context.customizations);
+        const defaultProps = Customizations.getSettings(fields, scope, this.context.customizations);
 
-              // tslint:disable-next-line:no-any
-              const componentProps = this.props as any;
+        // tslint:disable-next-line:no-any
+        const componentProps = this.props as any;
 
-              if (concatStyles) {
-                const mergedStyles = concatStyleSets(defaultProps.styles, componentProps.styles);
-                return <ComposedComponent {...defaultProps} {...componentProps} styles={mergedStyles} />;
-              }
+        if (concatStyles) {
+          const mergedStyles = concatStyleSets(defaultProps.styles, componentProps.styles);
+          return <ComposedComponent {...defaultProps} {...componentProps} styles={mergedStyles} />;
+        }
 
-              return <ComposedComponent {...defaultProps} {...componentProps} />;
-            }}
-          </CustomizerContext.Consumer>
-        );
+        return <ComposedComponent {...defaultProps} {...componentProps} />;
       }
 
       private _onSettingChanged(): void {